fix bug

+++ app/controllers/orders_controller.rb
@@ -489,7 +489,7 @@ class OrdersController < ApplicationController
         if !params_f["keyclientorders.batch_no".to_sym].blank?
           selectorders=selectorders.where("keyclientorders.batch_no = ?", params_f["keyclientorders.batch_no".to_sym])
         end
-        
+
         if !params_f[:is_printed].blank?
           selectorders=selectorders.where(is_printed: (params_f[:is_printed][0].eql?('t') ? true : false))
         end
@@ -1086,7 +1086,9 @@ class OrdersController < ApplicationController
         else
           order.update! tracking_number: tracking_number, transport_type: transport_type, total_weight: total_weight, pingan_ordertime: pingan_ordertime, customer_unit: customer_unit, customer_name: customer_name, customer_address: customer_address, customer_postcode: customer_postcode, province: province, city: city, county: county, customer_tel: customer_tel, customer_phone: customer_phone, status: status, user_id: current_user.id, is_printed: is_printed
 
-          order.order_details.destroy_all
+          if !params[:business_select].blank?
+            order.order_details.destroy_all
+          end
         
           # ords[0] = order
           # if find_has_stock(ords,false).blank?
